update similar files, merge and refomat shared files
