+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/Server/SqlSer.cs
@@ -197,10 +197,7 @@ namespace Microsoft.Data.SqlClient.Server
             _maxSize = _normalizer.Size;
         }
 
-        public override void Serialize(Stream s, object o)
-        {
-            _normalizer.NormalizeTopObject(o, s);
-        }
+        public override void Serialize(Stream s, object o) => _normalizer.NormalizeTopObject(o, s);
 
         public override object Deserialize(Stream s) => _normalizer.DeNormalizeTopObject(_type, s);
     }
@@ -253,22 +250,13 @@ namespace Microsoft.Data.SqlClient.Server
 
         public override long Position
         {
-            get
-            {
-                return _size;
-            }
-            set
-            {
-                _size = value;
-            }
+            get => _size;
+            set =>_size = value;
         }
 
         public override long Length => _size;
 
-        public override void SetLength(long value)
-        {
-            _size = value;
-        }
+        public override void SetLength(long value) => _size = value;
 
         public override long Seek(long value, SeekOrigin loc)
         {
