+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/Server/ValueUtilsSmi.cs
@@ -2729,7 +2729,6 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-
         //
         //  Semantics support routines
         //
