+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/SqlUserDefinedAggregateAttribute.cs
@@ -7,7 +7,7 @@ using Microsoft.Data.Common;
 
 namespace Microsoft.Data.SqlClient.Server
 {
-    /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/SqlUserDefinedAggregateAttribute/*' />
+    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/SqlUserDefinedAggregateAttribute/*' />
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = false, Inherited = false)]
     public sealed class SqlUserDefinedAggregateAttribute : Attribute
     {
@@ -19,11 +19,11 @@ namespace Microsoft.Data.SqlClient.Server
         private Format _format;
         private string _name;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/MaxByteSizeValue/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/MaxByteSizeValue/*' />
         // The maximum value for the maxbytesize field, in bytes.
         public const int MaxByteSizeValue = 8000;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/ctor/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/ctor/*' />
         // A required attribute on all UD Aggs, used to indicate that the
         // given type is a UD Agg, and its storage format.
         public SqlUserDefinedAggregateAttribute(Format format)
@@ -41,7 +41,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/MaxByteSize/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/MaxByteSize/*' />
         // The maximum size of this instance, in bytes. Does not have to be
         // specified for Native format serialization. The maximum value
         // for this property is specified by MaxByteSizeValue.
@@ -62,7 +62,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToDuplicates/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToDuplicates/*' />
         public bool IsInvariantToDuplicates
         {
             get
@@ -75,7 +75,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToNulls/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToNulls/*' />
         public bool IsInvariantToNulls
         {
             get
@@ -88,7 +88,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToOrder/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsInvariantToOrder/*' />
         public bool IsInvariantToOrder
         {
             get
@@ -101,7 +101,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsNullIfEmpty/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/IsNullIfEmpty/*' />
         public bool IsNullIfEmpty
         {
             get
@@ -114,11 +114,11 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/Format/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/Format/*' />
         // The on-disk format for this type.
         public Format Format => _format;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/Name/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedAggregateAttribute.xml' path='docs/members[@name="SqlUserDefinedAggregateAttribute"]/Name/*' />
         public string Name
         {
             get
